Prevent watch running more than once for multiple changes
When a bunch of files are changed at once, queue up the changes. There
are actually 2 queues to make for a nicer experience:

The first is a
250ms timeout so that it can capture a bunch of changes that happen at
once, for example as when happens when changing branch.

The second is a re-entrancy check so that if changes occur while it is
currently compiling a new compile will be deferred until it finishes (on
the same timeout timer).